Выполнено задание 16. Личный проект: разделяй и властвуй (Часть 1)
