+++ src/utils/common.js
@@ -1,30 +1,10 @@
-import {MONTH_NAMES} from "./const.js";
+import {MONTH_NAMES} from "../const.js";
 
 const MILLISECONDS_IN_SECUND = 1000;
 const SECUNDS_IN_MINUTE = 60;
 const MINUTES_IN_HOUR = 60;
 const HOURS_IN_DAY = 24;
 const DAYS = 30;
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  AFTEREND: `afterend`
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
 
 export const uniqueItems = (arr) => {
   return Array.from(new Set(arr));
@@ -113,10 +93,4 @@ export const getDatesDuration = (date1InInt, date2InInt) => {
     hours,
     minutes,
   });
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
 };
